Revert "Code review changes"
