+++ public/index.html
@@ -9,7 +9,7 @@
   <link href="https://fonts.googleapis.com/css?family=PT+Sans:400,700" rel="stylesheet">
 </head>
 <body>
-  <h1>FAC Skillshare </h1>
+  <h1>FaC Skillshare </h1>
   <form action="/submit" method="POST">
     <div id="container-form">
     <input type="text" name="name" placeholder="Your Name" required autocomplete="off">
@@ -25,8 +25,8 @@
     </div>
   </form>
   <div class="tab">
-        <button class="tablinks offerBtn" onclick="openTab(event, 'Offers')">Offers</button>
-        <button class="tablinks requestBtn" onclick="openTab(event, 'Requests')">Requests</button>
+        <button class="tablinks active" onclick="openTab(event, 'Offers')">Offers</button>
+        <button class="tablinks" onclick="openTab(event, 'Requests')">Requests</button>
     </div>
   <div class="offers tabcontent" id="Offers">
       <table id="offers-table">
