+++ src/postQueries.js
@@ -52,8 +52,7 @@ const updateDB = (name,email,skill,option,callback,cb) => {
     dbconnection.query(query, (err, insertOffer) => {
       if (err) {
         console.log(err);
-        throw new Error(err, 'the offer could not be added to offer table');
-        callback();
+        throw err;
       } 
       else { 
         // console.log('offer added sucessfully');
@@ -71,8 +70,7 @@ const updateDB = (name,email,skill,option,callback,cb) => {
     dbconnection.query(query, (err, insertOffer) => {
       if (err) {
         console.log(err);
-        throw new Error(err, 'the request could not be added to request table');
-        callback();
+        throw err;
       } 
       else {
         // console.log('request added sucessfully');
@@ -93,4 +91,4 @@ const addToDatabase = (name,email,skill,option, callback) => {
   });
 };
 
-module.exports = { addToDatabase, checkUsers, checkSkill, updateDB };
+module.exports = addToDatabase;
